Adding a remove-old-ipc option to start.
The idea being to allow the naught process to clean up after an erroneous
shutdown. Rather than having to manually run a check before starting the
naught process, which is more clumsy and less reliable (to my mind),
instead add a flag to naught which will make it try to unlink an
existing ipc file if an invalid one is found.